+++ changelog.md
@@ -1,5 +1,35 @@
 # PassMax Changelog
 
+0.2.2:
+
+- Simplify main loop, general prettification by @ethnh
+- Change license to Apache 2.0 for brew compatibility.
+
+0.2.1:
+
+- Add a master password
+- Small ui changes
+
+
+0.2.0 Octopus officer:
+
+- FULL ENCRYPTION
+- Ui fixes, including clear command
+- General fixes to my grammar!
+- More time!
+- General fixes to command function
+
+0.1.3:
+
+- Add export password
+- Add export all passwords (Of course encrypted)
+- Fix Bugs in encryption
+
+0.1.2:
+
+- Add encryption
+- Removed unnecessary dependencies
+
 0.1.1: 
 
 - Add importing